+++ doc/source/history.rst
@@ -8,6 +8,8 @@ Release history
 
 * ADDED: struct dtypes support in :py:func:`~reikna.cluda.dtypes.c_constant`.
 
+* ADDED: :py:func:`~reikna.cluda.dtypes.flatten_dtype` helper function.
+
 * CHANGED: added ``transposed_a`` and ``transposed_b`` keyword parameters to :py:class:`~reikna.matrixmul.MatrixMul`.
 
 * CHANGED: :py:class:`~reikna.core.transformation.KernelParameter` is not derived from :py:class:`~reikna.core.Type` anymore (although it still retains the corresponding attributes).